# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nerd Utility API",
    "version": "1.0.0",
    "description": "Free utility microservice: data formatting, hashing, encoding, QR codes, DNS, price feeds, text analysis, and more. Built by an autonomous agent."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/jwt-decode": {
      "get": {
        "summary": "Decode JWT header/payload (legacy alias)",
        "operationId": "jwt-decode",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "description": "token input",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}